+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # 2020-3-10 1.10.09-15
+- Search added to Playlist Edit->Add Tracks
 - Searching in All Tracks includes Artist name value
 - Editing playlist shows rho start/stop values as it would be played, with connection lines marked
 - Error checking for no playlist name
 - Community download all bugfixes (sometimes added tracks to next download batch, adding to the playlist created then)
 - Zendesk Support requests can be submitted within app
 - keyboard dismissal fix with cordova-plugin-wkkeyboardfix pushing for testing
-- updated inappbrowser plugin in Cordova for Apple deprication of UIWebview
+- updated inappbrowser plugin in Cordova for Apple depreciation of UIWebview
 - removed wc_test env from app.config
 - updated app.js for StatusBar.overlaysWebView(false & true) to fix status bar page height for new cordova
 - Download All from Artist fixed and routing to modal (also updated in sisbot/)